// src/setting/Types.ts
export type Category = {
  id: number;
  title: string;
  subCategories: SubCategory[];
  filters: Filter[];
};
type ThirdCategory = {
  id: number;
  title: string;
};

type CatalogSubCategory = {
  id: number;
  title: string;
  third_categories: ThirdCategory[];
};

export type CatalogCategory = {
  id: number;
  image: string;
  title: string;
  subCategories: CatalogSubCategory[];
};
export type SubCategory = {
  id: number;
  title: string;
  third_categories: { id: number; title: string }[];
};

export type Filter = {
  id: number;
  title: string;
  options: Option[];
};

type Option = {
  id: number;
  title: string;
  color_code: string | null;
};
export type HomeHero = {
  id: number;
  title: string;
  description: string;
  image: string;
};
export type Advanteges = {
  id: number;
  title: string;
  icon: string;
};
export type ProductResponse = {
  data: Product[];
  count: number;
  meta: Meta;
};

export type Product = {
  id: number;
  is_new: boolean;
  title: string;
  price: string;
  discount: string | null;
  discounted_price: string;
  unit: string | null;
  category: Category;
  sub_category: SubCategory;
  filters: {
    filter_id: number;
    filter_name: string;
    options: {
      option_id: number;
      name: string;
      is_default: string;
      color_code: string | null;
    }[];
  }[];
  brand: Brand;
  image: string;
  sliders: Slider[];
  slug: {
    en: string;
    ru: string;
  };
};

export type Brand = {
  id: number;
  title: string;
  logo: string;
};

type Slider = {
  id: number;
  image: string;
};

type Meta = {
  current_page: number;
  last_page: number;
  per_page: number;
  total: number;
};
export type RulesType = {
  id: number;
  title: string;
  description: string; // Since the description contains HTML-like content, you can use `string`.
};
export type Translation = {
  [key: string]: string;
};

export type TranslationsKeys = Record<keyof Translation, string>;
// ...
type BasketItem = {
  id: number;
  quantity: number;
  status: string;
  price: string;
  options: {
    filter: string;
    option: string;
  }[];
  product: {
    id: number;
    sub_category_id: number;
    category_id: number;
    title: string;
    short_title: string;
    is_new: boolean;
    is_stock: boolean;
    is_season: boolean;
    code: string;
    description: string;
    slug: {
      en: string;
      ru: string;
    };
    price: string;
    discount: number | null;
    discounted_price: string;
    comments_count: number | null;
    avg_star: number;
    unit: string | null;
    category: {
      id: number;
      title: string;
      subCategories: {
        id: number;
        title: string;
      }[];
      filters: {
        id: number;
        title: string;
        options: {
          id: number;
          title: string;
          color_code: string | null;
        }[];
      }[];
    };
    sub_category: {
      id: number;
      title: string;
    };
    brand: {
      id: number;
      title: string;
      logo: string;
    };
    image: string;
    sliders: {
      id: number;
      image: string;
    }[];
    comments: {
      id: number;
      comment: string | null;
      star: number;
      date: string;
      customer: {
        id: number;
        name: string;
        email: string;
        phone: string;
      };
    }[];
    options: {
      id: number;
      is_default: number;
      title: string;
      color_code: string | null;
    }[];
    rating_summary: string[];
  };
};

export type Basket = {
  basket_items: BasketItem[];
  total_price: number;
  discount: number;
  delivered_price: number;
  final_price: number;
};
// ...
export type Favorite = {
  id: number;
  product: Product;
};